Module Basics

Miscellaneous helpers.

Print helpers

module PPrt : sig ... end

Printable collections

module type ORDERED_TYPE = sig ... end

The signature of the collection builders below.

module type SET = sig ... end

The signature for sets, extended with printers.

module type MAP = sig ... end

The signature for maps, extended with printers.

module MakeSet (E : ORDERED_TYPE) : SET with type elt = E.t

Builds a printable set module.

module MakeMap (E : ORDERED_TYPE) : MAP with type key = E.t

Builds a printable map module.

module String : sig ... end

The Stdlib String module, extended with a printer and a hash function.

module Strings : SET with type elt = String.t

Printable string set module.

module StrMap : MAP with type key = String.t

Printable string map module.

module Integer : sig ... end

module Ints : SET with type elt = Integer.t

Printable integer set module.

module IntMap : MAP with type key = Integer.t

Printable integer map module.

module Digest : sig ... end

module Digests : SET with type elt = Digest.t

Printable digest set module.

val cmp_pair : ('a -> 'b -> int) -> ('c -> 'd -> int) -> ('a * 'c) -> ('b * 'd) -> int

Compare two pairs, with a priority on the first element.

Miscellaneous helpers

module SeqUtils : sig ... end

Memoization

val memoize : (module Stdlib.Hashtbl.HashedType with type t = 'k) -> ?init_size:int -> ('k -> 'a) -> 'k -> 'a

memoize (module K) ?init_size f

Sets up a memoization scheme on f by using a weak hash table to store its arguments (which must be of type K.t).

val memoize_poly : ?init_size:int -> ('k -> 'a) -> 'k -> 'a

Polymorphic memoization that relies on Stdlib.compare and Hashtbl.hash. Note the former has restrictions on the structures it may be safely applied to (it may raise Invalid_argument on functional values, or not terminate on cyclic structures).