Module Basics.PPrt

Useful aliases for printers

type pu = Stdlib.Format.formatter -> unit

type 'a pp = Stdlib.Format.formatter -> 'a -> unit

type 'a fmt = ('a, Stdlib.Format.formatter, unit) Stdlib.format

type ufmt = unit fmt

type ('a, 'b) func = ('a, Stdlib.Format.formatter, unit, 'b) Stdlib.format4 -> 'a

type 'a proc = ('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a

Printers

val pp_lst : ?fopen:ufmt -> ?fsep:ufmt -> ?fclose:ufmt -> ?fempty:ufmt -> 'a pp -> 'a list pp

pp_lst ~fopen ~fsep ~fclose ~fempty pp_e Returns a list printer with a custom open delimiter (fopen), a custom close delimiter (fclose) and a custom separator (fsep). If the list is empty, prints (fempty).

val pp_lst_i : ?fopen:ufmt -> ?fsep:ufmt -> ?fclose:ufmt -> ?fempty:ufmt -> ('a * int) pp -> 'a list pp

Same as pl, but with a counter incremented for each element of the list.

val pp_lstset : ?fsep:ufmt -> 'a pp -> 'a list pp

Same as pl, but prints the list with a set-like notation (lt1, elt2, ...)

val pp_arr : ?fopen:ufmt -> ?fsep:ufmt -> ?fclose:ufmt -> ?fempty:ufmt -> 'a pp -> 'a array pp

Same as pl, but for arrays.

val pp_arr_i : ?fopen:ufmt -> ?fsep:ufmt -> ?fclose:ufmt -> ?fempty:ufmt -> ('a * int) pp -> 'a array pp

Same as pli, but for arrays.

val pp_nodelim : ?fsep:ufmt -> ?fterm:ufmt -> 'a pp -> 'a list pp

Same as pl, but without an opening delimiter.

val pp_nodelim_i : ?fsep:ufmt -> ?fterm:ufmt -> ('a * int) pp -> 'a list pp

Same as pp_no_delim, but with a counter.

Helpers for printing string lists

module Strings : sig ... end

Helpers for pretty printing simple data structures

module UFmt : sig ... end

Miscellaneous printing helpers

val pp_time : Stdlib.Format.formatter -> Unix.tm -> unit

Prints the time with the format year-month-day-hour-min-sec.

val asprintf : ('a, Stdlib.Format.formatter, unit, unit, unit, Stdlib.Format.tag) Stdlib.format6 -> 'a

Version of Format.asprintf with virtually no right margin.

val kasprintf : (Stdlib.Format.tag -> 'a) -> ('b, Stdlib.Format.formatter, unit, unit, unit, 'a) Stdlib.format6 -> 'b

Version of Format.kasprintf with virtually no right margin.

val init_formatters : ?style_renderer:Fmt.style_renderer -> ?utf_8:bool -> unit -> unit

Setups the formatters with the correct margins (defined with the environment variable.